added task-1 and task-2
