+++ src/test-3/operationsTask-1.h
@@ -1,0 +1,11 @@
+#pragma once
+
+void reverse(int arr[], int len);
+
+void average(int arr[], int len);
+
+int compare(const void* a, const void* b);
+
+void fill_random(int arr[], int len);
+
+void printArr(int arr[], int len);
